+++ src/dsemantic.c
@@ -667,6 +667,12 @@ static void add_property_Variable(Sheet *sheet, size_t lineNum,
                         if (defaultValueToken->type == TK_FLOATLITERAL) {
                             newVariable.defaultValue.floatValue =
                                 defaultValueToken->data.floatValue;
+                        } else if (defaultValueToken->type ==
+                                   TK_INTEGERLITERAL) {
+                            // If we've been given an integer, that's fine,
+                            // we can easily convert it.
+                            newVariable.defaultValue.floatValue =
+                                (dfloat)defaultValueToken->data.integerValue;
                         } else {
                             d_error_compiler_push("Set variable to Float, got "
                                                   "non-float default value",
@@ -909,6 +915,12 @@ static void add_property_FunctionInput(Sheet *sheet, size_t lineNum,
 
                     if ((typeOfLiteral & argType) != TYPE_NONE) {
                         // The data types match, we can set the default value.
+                        defaultValue = literal->data;
+                    } else if (typeOfLiteral == TYPE_INT &&
+                               argType == TYPE_FLOAT) {
+                        // We can easily convert an integer into a float.
+                        literal->data.floatValue =
+                            (dfloat)literal->data.integerValue;
                         defaultValue = literal->data;
                     } else {
                         ERROR_COMPILER(
@@ -1414,6 +1426,18 @@ void d_semantic_scan_nodes(Sheet *sheet, SyntaxNode *root) {
 
                                             // Just set the LexData instead of
                                             // testing for each data type.
+                                            inputSocket->defaultValue =
+                                                literalToken->data;
+                                        } else if (typeOfLiteral == TYPE_INT &&
+                                                   inputSocket->type ==
+                                                       TYPE_FLOAT) {
+                                            // If the compiler expected a float,
+                                            // but got an integer, that's fine.
+                                            // We can easily convert it.
+                                            literalToken->data.floatValue =
+                                                (dfloat)literalToken->data
+                                                    .integerValue;
+
                                             inputSocket->defaultValue =
                                                 literalToken->data;
                                         } else {
@@ -2144,8 +2168,14 @@ void d_semantic_reduce_types(Sheet *sheet) {
                                             // it has already been set, then
                                             // we need to check our type is
                                             // the type that is already there.
-                                            if (inputType != TYPE_NONE && socket->type != inputType) {
-                                                d_error_compiler_push("Value inputs in a Ternary operator must be of the same type", sheet->filePath, node->lineNum, true);
+                                            if (inputType != TYPE_NONE &&
+                                                socket->type != inputType) {
+                                                d_error_compiler_push(
+                                                    "Value inputs in a Ternary "
+                                                    "operator must be of the "
+                                                    "same type",
+                                                    sheet->filePath,
+                                                    node->lineNum, true);
                                                 inputsSameType = false;
                                             }
 
@@ -2154,10 +2184,12 @@ void d_semantic_reduce_types(Sheet *sheet) {
 
                                         // So we need to check the connection.
                                         else if (socket->numConnections == 1) {
-                                            SheetSocket *otherSide = socket->connections[0];
+                                            SheetSocket *otherSide =
+                                                socket->connections[0];
 
                                             // Is this socket reduced?
-                                            if (IS_TYPE_REDUCED(otherSide->type)) {
+                                            if (IS_TYPE_REDUCED(
+                                                    otherSide->type)) {
 
                                                 // Great! So we can set this
                                                 // socket's type to be discrete!
@@ -2165,8 +2197,14 @@ void d_semantic_reduce_types(Sheet *sheet) {
 
                                                 // But wait... is is the same
                                                 // as the rest of the types???
-                                                if (inputType != TYPE_NONE && socket->type != inputType) {
-                                                    d_error_compiler_push("Value inputs in a Ternary operator must be of the same type", sheet->filePath, node->lineNum, true);
+                                                if (inputType != TYPE_NONE &&
+                                                    socket->type != inputType) {
+                                                    d_error_compiler_push(
+                                                        "Value inputs in a "
+                                                        "Ternary operator must "
+                                                        "be of the same type",
+                                                        sheet->filePath,
+                                                        node->lineNum, true);
                                                     inputsSameType = false;
                                                 }